requirements.txt
python-loki
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
